FOO-10: review fix v2

+++ src/main/java/com/ognevoydev/quisell/controller/PostController.java
@@ -1,16 +1,14 @@
 package com.ognevoydev.quisell.controller;
 
-import com.ognevoydev.quisell.common.exception.ForbiddenException;
-import com.ognevoydev.quisell.common.exception.NotFoundException;
+import com.ognevoydev.quisell.common.exception.HttpStatusException;
 import com.ognevoydev.quisell.model.Post;
 import com.ognevoydev.quisell.service.PostService;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.annotations.Where;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.web.bind.annotation.*;
 import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
+import static org.springframework.http.HttpStatus.FORBIDDEN;
 
 @RequiredArgsConstructor
 @RestController
@@ -36,13 +34,13 @@ public class PostController {
 
     @DeleteMapping("/{id}")
     public void deletePost(@PathVariable(value = "id") UUID postId, Principal principal) {
-//      String postAccountId = postService.getPostById(postId).getAccountId().toString();
-//        if(principal.toString().equals(postAccountId)){
+      String postAccountId = postService.findAccountIdById(postId).toString();
+        if(principal != null && principal.toString().equals(postAccountId)){
             postService.deletePostById(postId);
-//        }
-//        else {
-//            throw new ForbiddenException(postId, Post.class);
-//        }
+        }
+        else {
+            throw new HttpStatusException("Access to this resource on the server is denied", FORBIDDEN);
+        }
     }
 
 }
